+++ packages/client-config/src/client-config-writer/client_config_writer.ts
@@ -1,4 +1,4 @@
-import fs from 'fs/promises';
+import fsp from 'fs/promises';
 import { ClientConfig } from '../client-config-types/client_config.js';
 import path from 'path';
 import * as os from 'os';
@@ -16,17 +16,18 @@ export class ClientConfigWriter {
   ): Promise<void> => {
     const fileExtension = path.parse(targetPath).ext;
     switch (fileExtension) {
+      case '.ts':
       case '.js': {
         const fileContent = `export default ${JSON.stringify(
           clientConfig,
           null,
           2
         )}${os.EOL}`;
-        await fs.writeFile(targetPath, fileContent);
+        await fsp.writeFile(targetPath, fileContent);
         break;
       }
       case '.json':
-        await fs.writeFile(targetPath, JSON.stringify(clientConfig, null, 2));
+        await fsp.writeFile(targetPath, JSON.stringify(clientConfig, null, 2));
         break;
       default:
         throw new Error(