+++ packages/integration-tests/test-projects/data-storage-auth-with-triggers/amplify/auth.ts
@@ -1,7 +1,11 @@
 import { Auth } from '@aws-amplify/backend-auth';
+import { myFunc } from './function.js';
 
 export const auth = new Auth({
   loginWith: {
     email: true,
+  },
+  triggers: {
+    postConfirmation: myFunc,
   },
 });