+++ packages/plugin-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aws-amplify/plugin-types
 
+## 0.1.1-alpha.4
+
+### Patch Changes
+
+- 3c36ace: Implement UserPool trigger config
+
 ## 0.1.1-alpha.3
 
 ### Patch Changes